fix: use the right netstat path on windows
The program using this package can be in any of the drives on windows machine. The package expected the user of this package to be in C: and did not include the drive path in command. Update the command to use `which`, to compute the right path for netstat command